Added Timestamp to filenames
If previous version of story exists, do not overwrite. THis is done because sometimes I have made edits. Also, sometimes authors claim to have rewritten stuff and it would be interesting to see how much has changed.

This is implemented by adding a timestamp to the filenames if a previous version exists.. This means there could be many versions but, hopefully, on seeing the timestamp, I would investigate why there are multiple copies and delete the un-needed ones.