Attempted Improvements to Image Retrieval
Sometimes pictures have to be downloaded manually. Made changes to not delete downloaded pictures so they can be re-used when a story is rebuilt.

Also, sometime HTML is returned instead of an image. Will now check for a different address in the new docuemnt. (So far only case returned same document requiring a manual download).